backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
boto3==1.16.4
botocore==1.19.4
certifi==2019.11.28
cycler==0.10.0
distlib==0.3.0
Django==3.1.2
django-rename-app==0.1.2
django-storages==1.10.1
djangorestframework==3.12.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.2
jmespath==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.17.2
pathlib==1.0.1
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.1
urllib3==1.25.11
virtualenv==20.0.14
virtualenv-clone==0.5.3
whitenoise==5.2.0
zipp==3.1.0
